web/backend/requirements.txt
django
djangorestframework

django-filter

django-cors-headers
django_extensions
psycopg2

dj-database-url
djangorestframework-simplejwt
Pillow
drf-yasg
